+++ src/common/components/DefaultProfileImage/utils/hash-code-utils.ts
